Refactored some UINT to uint64 This has become necessary as for some large upload the session upload was overflowing. Not causing issues per se, but..

+++ srchybrid/Emule.cpp
@@ -807,13 +807,13 @@ BOOL CALLBACK CemuleApp::SearchEmuleWindow(HWND hWnd, LPARAM lParam) noexcept
 }
 
 
-void CemuleApp::UpdateReceivedBytes(uint32 bytesToAdd)
+void CemuleApp::UpdateReceivedBytes(uint64 bytesToAdd)
 {
 	SetTimeOnTransfer();
 	theStats.sessionReceivedBytes += bytesToAdd;
 }
 
-void CemuleApp::UpdateSentBytes(uint32 bytesToAdd, bool sentToFriend)
+void CemuleApp::UpdateSentBytes(uint64 bytesToAdd, bool sentToFriend)
 {
 	SetTimeOnTransfer();
 	theStats.sessionSentBytes += bytesToAdd;
